pyproject.toml
[project]
name = "harbinger"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
